requirements.txt
Django
python-dotenv
openpyxl
fpdf2
groq
twilio
requests
